Updated the workflow search and add to watch list
If the search result contains more than one item, we will display the
list of items in the search result and user will need to add them by
clicking on the button. If the search result only contains one item, the
item is added to the watch list and the search result window is closed.