Added Investment page
Investment is to aid players to flip the market. User inputs the price
and quantity they bought an item, and we calculate with the current
market offering price, what is the gain/lost the user will be hit with.
This page is similar to item search page, it will require a popup for
the user to input the amount and quantity that the item was bought.